updating helper and config files, removed app.py and made local_app.py work with static data, reddit working correctly
