+++ local_app.py
@@ -1,56 +1,67 @@
-from flask import Flask
-server = Flask('my app')
+#from flask import Flask
+#server = Flask('my app')
 
-import dash
-import dash_table
-import pandas as pd
+# Standard imports
 import datetime
+import pandas as pd
 import pytz
 import json
 import os
+import twitter
 from td.client import TDClient
-from config import account_number, client_id
-import td_helper
-import reddit_helper
+
+# Dash imports
+import dash
+import dash_table
 import dash_html_components as html
 import dash_core_components as dcc
-from dash.dependencies import Input, Output
 import plotly.graph_objects as go
-import twitter
+from dash.dependencies import Input, Output
+
+# Helper and config imports
+import td_config
+import td_helper
+import reddit_config
+import reddit_helper
 import twitter_config
+import twitter_helper
+
+# Static data for testing
+import static_data
 
 # Setup table
-symbols = td_helper.get_symbols()
-df_quotes = td_helper.get_radar_quotes(symbols)
+symbols = static_data.static_symbols #td_helper.get_symbols()
+df_quotes = pd.read_json(static_data.static_quotes) #td_helper.get_radar_quotes(symbols)
 
 # Setup tweets table
 try:
-    df_tweets = td_helper.fetch_tweets(tweet_symbols)
+    df_tweets = pd.read_json(static_data.static_tweets) #td_helper.fetch_tweets(tweet_symbols)
 except:
     df_tweets = pd.DataFrame({'Date': ['Please wait'], 'User': ['Twitter API rate limit'], 'Tweet': ['currently exceeded.']})
 
 # # Setup reddit table
 try:
-    df_reddit = reddit_helper.get_reddit()
+    df_reddit = reddit_helper.get_reddit(reddit_config.client_id, reddit_config.client_secret, reddit_config.user_agent)
 except:
     df_reddit = pd.DataFrame({'Date': ['Please wait'], 'User': ['Reddit API rate limit'], 'Tweet': ['currently exceeded.']})
+
 # Initial chart one
-volatile_tickers = df_quotes['Symbol'][:2]
-df_p_hist = td_helper.get_radar_prices(1,5,volatile_tickers)
-df_cs_one = df_p_hist.query("Symbol == @volatile_tickers[0]").reset_index(drop=True)
+#volatile_tickers = df_quotes['Symbol'][:2]
+#df_p_hist = td_helper.get_radar_prices(1,5,volatile_tickers)
+df_cs_one = pd.read_json(static_data.static_candle) #df_p_hist.query("Symbol == @volatile_tickers[0]").reset_index(drop=True)
 
 fig_cs_one = go.Figure(go.Candlestick(x=df_cs_one['Date'], open=df_cs_one['Open'], high=df_cs_one['High'],
                                       low=df_cs_one['Low'], close=df_cs_one['Close']))
 
 fig_cs_one.update_layout(paper_bgcolor='rgba(0,0,0,0)', plot_bgcolor='rgba(0,0,0,0)', font_color='white',
-                         title_text='{} 5-minute Chart'.format(volatile_tickers[0]), title_font_size=20,
+                         title_text='{} 5-minute Chart'.format('NIO'), title_font_size = 20, #volatile_tickers[0]), title_font_size=20,
                          xaxis = dict(rangeslider = dict(visible = False)))
 
 fig_cs_one.update_xaxes(gridcolor = '#D2D2D2')
 
 
 ## App
-app = dash.Dash('Dash Hello World', external_stylesheets=['https://codepen.io/chriddyp/pen/bWLwgP.css'], server = server)
+app = dash.Dash('Dash Hello World', external_stylesheets=['https://codepen.io/chriddyp/pen/bWLwgP.css']) #, server = server)
 
 text_style = dict(color='#444', fontFamily='sans-serif', fontWeight=300)
 
@@ -86,52 +97,50 @@ app.layout = html.Div([
         style_header = {'backgroundColor': '#a1a1a1', 'color': 'black'}, style_cell = {'textAlign': 'center', 'font_size': '18px', 'white-space': 'normal','word-wrap': 'break-word',}), className = 'six columns'),
 ], style = {'margin-top': '50px', 'margin-left': '50px'})
 
-@app.callback(Output('transport-value', 'children'), [Input('interval-component', 'n_intervals')])
-def update_symbols(n):
-    symbols = td_helper.get_symbols()
-    return json.dumps(symbols)
+#@app.callback(Output('transport-value', 'children'), [Input('interval-component', 'n_intervals')])
+#def update_symbols(n):
+#    symbols = td_helper.get_symbols()
+#    return json.dumps(symbols)
 
-@app.callback(Output('table', 'data'), [Input('transport-value', 'children')])
-def update_table(j_symbols):
-    df_quotes = td_helper.get_radar_quotes(json.loads(j_symbols)).to_dict('records')
-    return df_quotes
+#@app.callback(Output('table', 'data'), [Input('transport-value', 'children')])
+#def update_table(j_symbols):
+#    df_quotes = td_helper.get_radar_quotes(json.loads(j_symbols)).to_dict('records')
+#    return df_quotes
 
-@app.callback(Output('reddit-table', 'data'), [Input('transport-value', 'children')])
-def update_reddit():
-    try:
-        df_reddit = reddit_helper.get_reddit().to_dict('records')
-    except:
-        df_reddit = pd.DataFrame({'Date': ['Please wait'], 'User': ['Reddit API rate limit'], 'Tweet': ['currently exceeded.']}).to_dict('records')
-    return df_reddit
+#@app.callback(Output('reddit-table', 'data'), [Input('transport-value', 'children')])
+#def update_reddit():
+#    try:
+#        df_reddit = reddit_helper.get_reddit().to_dict('records')
+#    except:
+#        df_reddit = pd.DataFrame({'Date': ['Please wait'], 'User': ['Reddit API rate limit'], 'Tweet': ['currently exceeded.']}).to_dict('records')
+#    return df_reddit
 
-@app.callback(Output('tweet-table', 'data'), [Input('transport-value', 'children')])
-def update_reddit():
-    try:
-        df_tweets = td_helper.fetch_tweets(tweet_symbols).to_dict('records')
-    except:
-        df_tweets = pd.DataFrame({'Date': ['Please wait'], 'User': ['Twitter API rate limit'], 'Tweet': ['currently exceeded.']}).to_dict('records')
-    return df_tweets
+#@app.callback(Output('tweet-table', 'data'), [Input('transport-value', 'children')])
+#def update_reddit():
+#    try:
+#        df_tweets = td_helper.fetch_tweets(tweet_symbols).to_dict('records')
+#    except:
+#        df_tweets = pd.DataFrame({'Date': ['Please wait'], 'User': ['Twitter API rate limit'], 'Tweet': ['currently exceeded.']}).to_dict('records')
+#    return df_tweets
 
-@app.callback(Output("candle-one", "figure"), [Input('transport-value', 'children')])
-def display_candlestick(j_symbols):
-    volatile_tickers = td_helper.get_radar_quotes(json.loads(j_symbols))['Symbol'][:2]
-    df_p_hist = td_helper.get_radar_prices(1,5,volatile_tickers)
-    df_cs_one = df_p_hist.query("Symbol == @volatile_tickers[0]").reset_index(drop=True)
+#@app.callback(Output("candle-one", "figure"), [Input('transport-value', 'children')])
+#def display_candlestick(j_symbols):
+#    volatile_tickers = td_helper.get_radar_quotes(json.loads(j_symbols))['Symbol'][:2]
+#    df_p_hist = td_helper.get_radar_prices(1,5,volatile_tickers)
+#    df_cs_one = df_p_hist.query("Symbol == @volatile_tickers[0]").reset_index(drop=True)
 
-    fig_cs_one = go.Figure(go.Candlestick(x=df_cs_one['Date'], open=df_cs_one['Open'], high=df_cs_one['High'],
-                                          low=df_cs_one['Low'], close=df_cs_one['Close']))
+#    fig_cs_one = go.Figure(go.Candlestick(x=df_cs_one['Date'], open=df_cs_one['Open'], high=df_cs_one['High'],
+#                                          low=df_cs_one['Low'], close=df_cs_one['Close']))
 
-    fig_cs_one.update_layout(paper_bgcolor='rgba(0,0,0,0)', plot_bgcolor='rgba(0,0,0,0)', font_color='white',
-                             title_text='{} 5-minute Chart'.format(volatile_tickers[0]), title_font_size=20,
-                             xaxis = dict(rangeslider = dict(visible = False)))
+#    fig_cs_one.update_layout(paper_bgcolor='rgba(0,0,0,0)', plot_bgcolor='rgba(0,0,0,0)', font_color='white',
+#                             title_text='{} 5-minute Chart'.format(volatile_tickers[0]), title_font_size=20,
+#                             xaxis = dict(rangeslider = dict(visible = False)))
 
-    fig_cs_one.update_xaxes(gridcolor = '#D2D2D2')
+#    fig_cs_one.update_xaxes(gridcolor = '#D2D2D2')
 
-    return fig_cs_one
+#    return fig_cs_one
 
-app.css.append_css({'external_url': 'https://codepen.io/chriddyp/pen/bWLwgP.css'})
+#if __name__ == '__main__':
+#    app.server.run()
 
-if __name__ == '__main__':
-    app.server.run()
-
-#app.server.run(debug = True)
+app.server.run(debug = True)
