+++ static_data.py
@@ -1,0 +1,7 @@
+static_tweets = '{"Date":{"0":1611777925000,"1":1611777924000,"2":1611777923000,"3":1611777919000,"4":1611777917000,"5":1611777916000,"6":1611777915000,"7":1611777913000,"8":1611777910000,"9":1611777910000,"10":1611777902000,"11":1611777900000,"12":1611777891000,"13":1611777865000,"14":1611777860000,"15":1611777848000,"16":1611777845000,"17":1611777843000,"18":1611777836000,"19":1611777827000,"20":1611777792000,"21":1611777792000,"22":1611777791000,"23":1611777787000,"24":1611777772000,"25":1611777766000,"26":1611777727000,"27":1611777723000,"28":1611777718000,"29":1611777676000,"30":1611777673000,"31":1611777640000,"32":1611777606000,"33":1611777569000,"34":1611777562000,"35":1611777560000,"36":1611777396000,"37":1611777333000,"38":1611777177000,"39":1611777156000,"40":1611776993000,"41":1611776869000,"42":1611776855000,"43":1611776825000,"44":1611776815000,"45":1611776761000,"46":1611776647000,"47":1611776631000,"48":1611776561000,"49":1611776380000,"50":1611776274000,"51":1611776237000,"52":1611776235000,"53":1611776228000,"54":1611776133000,"55":1611776085000,"56":1611776057000,"57":1611776041000,"58":1611775912000,"59":1611775889000,"60":1611775378000,"61":1611775345000,"62":1611774570000,"63":1611774374000,"64":1611774314000,"65":1611774230000,"66":1611774100000,"67":1611774083000,"68":1611774018000,"69":1611774002000,"70":1611773633000,"71":1611773180000,"72":1611771996000,"73":1611771852000,"74":1611771783000,"75":1611771031000,"76":1611770923000,"77":1611770812000,"78":1611770491000,"79":1611769911000,"80":1611767341000,"81":1611767229000,"82":1611767117000,"83":1611767029000,"84":1611766953000,"85":1611754391000,"86":1611753051000,"87":1611746128000,"88":1611736106000,"89":1611721668000,"90":1611712182000,"91":1611707727000,"92":1611675506000,"93":1611665676000,"94":1611661422000,"95":1611642215000,"96":1611473153000,"97":1611417949000,"98":1611337508000,"99":1611277867000,"100":1611100185000},"User":{"0":"News89 Media","1":"Tara Ziegmont","2":"equality for all","3":"Together we did","4":"fullmetal birdchemist \\u270a\\ud83c\\udffc bIm","5":"Kim Persson","6":"\\u1d0d\\u1d00\\u0280\\u026a \\ud83d\\udc38","7":"Live Online","8":"chrissy \\ud83c\\udf3a\\u2728 \\/\\/ ACAB","9":"Steve Carter","10":"Jon Jefferson","11":"Collectibles & More","12":"Tim","13":"Kurisu\\u30af\\u30ea\\u30b9","14":"TheWealthHustle","15":"Market Rebellion","16":"Nick Ferrari","17":"\\u2601\\ufe0f\\ud83e\\udd78\\ud83e\\udd86C says abolish the EC","18":"sad saggy scrotum","19":"Raisa","20":"harvest bistro","21":"Tj - Stocks \\ud83d\\udcc8\\u2022 Options \\ud83d\\udcb0\\u2022 Forex \\ud83d\\udcb4","22":"Ruby4277","23":"Mama Mara","24":"Mara\\ud81a\\uddf5\\u2077\\u1d2e\\u1d31 ia","25":"David Dandaneau \\ud83e\\udd28","26":"#moneylatte","27":"harvest bistro","28":"J.","29":"\\ud83e\\udd85","30":"Jesse Joseph Knox","31":"Thomas James","32":"DK Nation","33":"g.18x277.777","34":"Anthony DePalma","35":"Johnny Cash","36":"naashonomics","37":"\\u6e80\\u5929 @ 10:10 for GoSe\\ud83e\\udd55\\ud83e\\udd55\\ud83e\\udd55","38":"Joe Schmoe","39":"thelonelytrader","40":"Anderson\'s Epic Adventures","41":"naashonomics","42":"Mike Moody","43":"WindmillCancerSurvivor","44":"Dennis Coyle","45":"AliveNhappy","46":"STL Cardinals","47":"Blazing Saddles","48":"naashonomics","49":"Stock Market Genius","50":"marc","51":"James","52":"Noirdos","53":"Sawgrass Bay Elem","54":"Navi Avatar \\ud83c\\uddfa\\ud83c\\uddf8","55":"Wayne Nguyen","56":"Thoreau","57":"Joe Duncan","58":"Crystal Castle Trading","59":"Craig Carmon","60":"Austin Moore-Farrow","61":"CachorroRoxo\\ud83c\\uddfa\\ud83c\\uddf2\\ud83d\\udc02","62":"\\u3148\\u3139","63":"Elizabeth W.","64":"sahba masoumi","65":"Liz Lele","66":"CAPM_TECHNICAL_ANALYSIS","67":"Dan\\ud83c\\udde7\\ud83c\\uddfb\\ud83c\\udde9\\ud83c\\uddf0 | Nortrade","68":"ankit malik","69":"Jollibee","70":"Branden","71":"Winns","72":"Blue Collar Dollar \\ud83d\\udde3\\ud83c\\uddfa\\ud83c\\uddf8","73":"Blue Collar Dollar \\ud83d\\udde3\\ud83c\\uddfa\\ud83c\\uddf8","74":"Alex","75":"Nicole","76":"Mr Divies","77":"United States Oil and Natural Gas","78":"United States Oil and Natural Gas","79":"Stephen","80":"Paul Cerro","81":"Dennis Coyle","82":"Simon\'s Trade","83":"Henrik H","84":"Bullitt Capital","85":"Mike McClure","86":"NIO","87":"Mark Minervini","88":"The Iron Sheik","89":"Belfast News Letter","90":"ProPublica","91":"Chamath Palihapitiya","92":"Mara Silvers","93":"Kgopolo Mphela","94":"Mara Silvers","95":"ProPublica","96":"Scott Redler","97":"Conor McGregor","98":"Chris Randone","99":"Gillian Tan","100":"Ross Gerber"},"Tweet":{"0":"Sbe Opening 50 New Hotels Around the World by\\u00a02022 https:\\/\\/t.co\\/SewrptUbMa","1":"@ElizabethNorton And, just so you are aware, saying what you are saying can get you banned from Twitter for spreadi\\u2026 https:\\/\\/t.co\\/KAbbEi3OVj","2":"@Mohit51195052 @saahilmenghani Jo kal huya vo galat tha but jo small farmers anpad hein or unko ab bhi darr hae ki\\u2026 https:\\/\\/t.co\\/EJcv4yuVhz","3":"For all us outsider traders - the wsb hedge fund is coming\\n#WSB $GME $BB $TSLA $PLTR $NIO $BBBY $KODK $NOK $AMC $AMD $SQ $CODX $EXPR $SRNE","4":"@rickenchikin it\'s kind of a shame bc the idea of someone impersonating phoenix now that he\'s gaining notoriety is\\u2026 https:\\/\\/t.co\\/xlpOh3nspo","5":"If #CCIV hits 25$ before the end of the week, I\\u2019ll giveaway 500$\\ud83d\\udcb8\\ud83d\\ude80\\nFollow me and retweet to enter\\ud83d\\udcc8\\n\\n#SPACs #SPAC\\u2026 https:\\/\\/t.co\\/xXBcAvNLIX","6":"@foreign_mami_li i will personally toe tag that man that slapped your thigh at work","7":"\\u26a1\\ufe0f \\u201cLive\'NOW:: 2nd Annual LI Winter Classic Women 2021\\u201d\\n\\nhttps:\\/\\/t.co\\/TbXh3gExdh","8":"@AnnieBecky also https:\\/\\/t.co\\/iBoapPzU5u","9":"@GOPLeader Says the Party of Do Nothing Trump. Hurry down to Mara-Largo I\\u2019m sure your man needs you to wipe his butt.","10":"Cover Reveal: Forever After All by Catharine Mara #ContemporaryRomance @EJBookPromos https:\\/\\/t.co\\/il53rn0Vcq via\\u2026 https:\\/\\/t.co\\/Q8fIfjXW0Q","11":"Check out what I just listed on Mercari. Tap the link to sign up and get up to $30 off. https:\\/\\/t.co\\/ng540r8aSM","12":"I made my bag, but I don\\u2019t mind making a second years salary tomorrow? We doing this? $AMC $PLTR https:\\/\\/t.co\\/XkX70fC5vA","13":"@blkgrlmagicc @D0ubl3St3P The CCIV merger is delayed","14":"@1charts6 Gday mate, from austrAlia. Jmia looks to want to reverse a bit, are you looking to accumulate?","15":"WSB Stocks After-Hours Update:\\n\\n$PLTR -5%\\n$NOK -9%\\n$BB -10%\\n$BBBY -12%\\n$GME -16%\\n$AMC -26%","16":"@Mara_E_Blake Lovely \\ud83d\\ude0d","17":"@t_scott89 What\\u2019s weirder is that Mara Wilson is their cousin (she has expressed disdain of them)","18":"@The_RockTrading takin some tips from nio! i like it","19":"@papihayson Lmaooo ima be back on my diet tomorrow. No worries I\\u2019m protected so it didn\\u2019t fall on the floor \\ud83d\\ude02\\ud83d\\ude2d","20":"@AlexDelarge6553 $cciv and $gik add $atvk and $opti if they are down. Same for $inqd $sndl","21":"If someone wants to pump $SBE I wouldn\\u2019t be apposed \\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude02","22":"$KULR....market took everything down...but see $TSLA? $NIO? Insert anything EV? This is the tech that will service the batteries....","23":"Halsey\\u2019s pregnant with her rainbow baby and i\\u2019m going to cry for the rest of my life i\\u2019m so happy for her omg \\ud83d\\ude2d\\ud83d\\ude2d\\ud83d\\ude2d","24":"Hell we didn\'t even get a studio verison of Tony Montana ft jimin \\ud83d\\udc94","25":"In other news... nibbled at this guy today \\u27a1\\ufe0f MGM Resorts-Backed @playstudios Is Discussing #SPAC Deal $ACAC $ACACW\\u2026 https:\\/\\/t.co\\/UEjyuqoIQ9","26":"who do you support ?\\n\\n$AMC $GME $CCIV $NIO $NOK $BB $IPOD $IPOE\\n@InvestorAshar  @ChrisCamillo  @miakhalifa","27":"@AlexDelarge6553 $cciv and $gik add $atvk and $opti if they are down. Same for $inqd","28":"@TheStephenChiu Haha dude same! Watch $CCIV and $FUBO tmr","29":"@CenturyOG @MrZackMorris @yatesinvesting @AlfPoor @AmandaCStocks @StanTradngman @buysellshortnet @elonmusk\\u2026 https:\\/\\/t.co\\/YNCApimqSy","30":"@Lovewynns Definitely not. Great news for $TSLA. Even though it appears that Lucid &amp; Tesla are competing, it\'s only\\u2026 https:\\/\\/t.co\\/sBIwT5XtoM","31":"Hope is a 4 letter word. \\nTrade accordingly. $CCIV","32":"GOLF BEST BETS: @reidtfowler provides his top PGA TOUR golf bets on @DKSportsbook for the Farmers Insurance Open.\\u2026 https:\\/\\/t.co\\/uxVuoQLuDx","33":"So y\\u2019all like stocks now? Boost $nndm for me","34":"Fun day in the stonk market today. Time to celebrate! $NIO $NNDM $AMC $NOK https:\\/\\/t.co\\/c45QEWKXVo","35":"ALARM SET TOMORROW IS GOING TO BE WILD $SNDL $AMC $NAKD $GME $NOK $CCIV $CLOV $CIIC $LOTZ $BNGO https:\\/\\/t.co\\/oDOHrO8qi6","36":"#cnbc for once said #WallStreet is not the decider for #stock  $GME $EXPR $AMC $NOK they were wrong on #bitcoin\\u2026 https:\\/\\/t.co\\/nv7goIbToF","37":"I\\u2019ve watched the junshua huya stream so many times I\\u2019m thinking I might as well sub it fjdkdkd","38":"@RevShark I went on office on offense with small buys in the early AM panic in DKNG and PEIX. Larger buy before the\\u2026 https:\\/\\/t.co\\/uA4snnb5UQ","39":"$CLOV $SBE $F $AMC $BB $NOK \\ud83d\\udc40\\ud83d\\ude80 https:\\/\\/t.co\\/8kh3rptU11","40":"Hold The Line Buy ANY! Dip $AMC $GME $BBBY $SKT $NOK $TLRY $M $SIRI \\ud83e\\udd11\\ud83d\\udca5HUNKER\\ud83d\\udca5DOWN IT\'S WAR! TIME GET YOUR LUNA BARS\\u2026 https:\\/\\/t.co\\/pCqkcwCgZt","41":"# billionare lost money it\'s  national issue for #sec when #citron called $JMIA fraud tanked stock from 55 to 5 it\\u2026 https:\\/\\/t.co\\/HX6tEz6KXu","42":"$NNDM whatever happens over next 5 trading days, don\\u2019t sell! it\\u2019ll snap right back to 15 imho \\ud83e\\udd19","43":"@DanielSpeight8 @BilbroMccaleb @charliekirk11 Tons of people did and still are. PLUG, TESLA, HYLN, FCEL, SBE.. the\\u2026 https:\\/\\/t.co\\/wML0MqrG7x","44":"The Stock Market Is Not a Game, It\'s serious Shit. I take it Serious and You should too! Long and Strong 27 Compani\\u2026 https:\\/\\/t.co\\/0v2Wah4oUL","45":"@CNBCnow That\\u2019s nonsense so we have to be slaves to some fuckkng pedophile billionaires. Why the fuck you didn\\u2019t ba\\u2026 https:\\/\\/t.co\\/vqHWKYhkfZ","46":"@franklyfacts @MrZackMorris @cappingbird Same to you.\\n\\nIf you\\u2019re a long-term hold type of person, I strongly recomm\\u2026 https:\\/\\/t.co\\/H450bj8gsD","47":"@mukund My overarching goal is long term unrealized capital gains with stocks that have good runways, secular growt\\u2026 https:\\/\\/t.co\\/3xhyQLiSvh","48":"when citron comes and @cnbc and calls $TSLA $JMIA fraud #SEC does nothing call it\'s #FreedomOfSpeech but\\u2026 https:\\/\\/t.co\\/qKwETTFBEB","49":"$JMIA Jumia Technologies AG Ame Top stock up 2423% from low. Close: 57.37 VolvsAvg: 0.88 Liq: $708M https:\\/\\/t.co\\/iEd8n93IpL","50":"@GetBenchmarkCo DKNG in gambling for sure","51":"10+ states about to legalize mobile sports betting this year, buy $DKNG instead of gamestop you morons","52":"@BlackXSupremacy I made good money today on JMIA, PLTR and NK.  Mmmm hmmm.","53":"Virtual Guest Reader Videos- hear some SBE teachers past and present read their favorite books. You will also hear\\u2026 https:\\/\\/t.co\\/2oFrB7kRE5","54":"Dow down 633, $AAPL $AMZN $TSLA $NFLX $DKNG $FUBO $FSLY $SPY $QQQ https:\\/\\/t.co\\/EifRzd04Vm","55":"@wyzzil As of yesterday they accumulated 11.267 mil shares of $NNDM https:\\/\\/t.co\\/moVg1Xyvbn","56":"So I already get the ARK daily newsletter but that doesn\\u2019t include EVERYTHING they buy daily. \\n\\nDoes anyone have an\\u2026 https:\\/\\/t.co\\/5diRX9ayP0","57":"@LincolnWDaniel I used it as an excuse to invest in indexes, ETFs, PINS, and of course your mentions, TTCF, SBE, and more.","58":"Todays Plays: Closed... $AMC, $TSLA CS, some $DBX leaps after the nice break out, $JMIA &amp; $SBE Calendars for some n\\u2026 https:\\/\\/t.co\\/Iqv80TOTCK","59":"@stockzdaddy @SilkLifeMedSpa 3.99 float. And smaller share structure than $BYND &amp; $TTCF. \\n\\n$NXMH will make big waves!","60":"$CLII $LI $WY $XPEV $FCEL if you want to invest in the environment!!\\nThat\\u2019s the real wave \\ud83c\\udf0a #savetheplanet #goddamnt","61":"Calls held:\\n2\\/19\\n$AMD $110 $AQB 12.5 $FB 310\\n$HEXO 7 $ITP 1 $MITT 4\\n$MKND 3 $OEG 10 $QLGN 5\\n$SNDL 1 $SOS 5 $TTCF 30\\u2026 https:\\/\\/t.co\\/4PWgzXdzgh","62":"[NEOSUBS] 190219 Huya Game Live Broadcast With Jaemin &amp; Renjun https:\\/\\/t.co\\/mkMH0FTf8R via @YouTube","63":"$NNDM ARK adds 109,590 shares today 1\\/27\\/21\\n\\n+85,632 shares to ARKQ \\nNo shares added to ARKW \\n\\n+23,958 shares to IZRL","64":"$Fsr THE FACT THAT FISKER IS THE ONLY GREEN STOCK RN. Wowwwwww\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\udfe2\\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude02\\ud83d\\ude1b\\ud83d\\udcc8\\ud83d\\udcc8\\ud83d\\udcc8 20 tomorrow. Get in. Fisker i\\u2026 https:\\/\\/t.co\\/RqHulATmw4","65":"$46.50 win on #GME long and #NNDM short\\n\\nVideo : https:\\/\\/t.co\\/XVhEiDX1Fq\\n\\nI\'m learning from:\\u2026 https:\\/\\/t.co\\/6mUnKUkVx7","66":"TESLA STOCK BEST BUYING STRATEGY AFTER HUGE Q4 EARNINGS MISS! TESLA STOCK NEWS &amp; ANALYSIS (TSLA)\\n\\n\\ud83d\\udea8\\u2026 https:\\/\\/t.co\\/sdpoukZNZG","67":"@squawksquare Look at $NNDM @CathieDWood has been loading up big time $ARKQ, $ARKW.","68":"@wyzzil I believe ARK added NNDM today as well 109,590 shares","69":"$TTCF I like what it did today despite of market environment. This is a strong ticker. I have a small size position\\u2026 https:\\/\\/t.co\\/z5eUPOIbJY","70":"@ddbuffett Congrats, I went fro August 60k to 145k Jan. Lets keep going! $NNDM","71":"plant food stocks $bynd 25% of float shorted $ttcf 33% $shak next short squeeze? $amc $gme $gsx $koss $fizz","72":"@cmarcucci6 And SPRQ, CLII , APSG, and DCRB all have new spacs either just IPO or about to. \\n\\nHere\\u2019s the current sp\\u2026 https:\\/\\/t.co\\/sllQsvUv80","73":"@cmarcucci6 I wouldn\\u2019t be exaggerating to say I have over 15,000 warrants with price averages near $2 or below lol.\\u2026 https:\\/\\/t.co\\/5ZzN5sKOUg","74":"Good signs for $TTCF https:\\/\\/t.co\\/vInfRRYrW0","75":"$TTCF 4 hr chart. Asc channel off bull run, consolidating w\\/ declining volume, hidden bull div in Stoch. Lower tren\\u2026 https:\\/\\/t.co\\/VXQdZd2ICb","76":"@evgonetwork @GM When is EVgo and CLII doing the RM","77":"Ford Electric Vehicles to be in Line with $TSLA Tesla competition $F \\n\\n$10 stock vs $800 - which would you rather g\\u2026 https:\\/\\/t.co\\/OdfbbTAaAN","78":"Electric Farming Tractors and Heavy Dump Trucks.. \\n\\n$IDEX \\n\\nIdeanomics prepares to launch its heavy truck and speci\\u2026 https:\\/\\/t.co\\/JXW7I6Umnf","79":"$ttcf #reddit make it happen \\ud83d\\ude4f","80":"Added $CLII and $CLOV today\\n\\nTotal positions:\\n$SPCE\\n$SSPK\\n$SRAC\\n$NPA\\n$CLOV\\n$CLII\\n$GTBIF\\n$DAL\\n$WKHS\\n$DKNG","81":"Only 234,000,000 Shares Outstanding for $NAKD I am Long these, $NIO $TSLA $AAPL $AMZN $XPEV $LI $RIDE $GEVO $ABML\\u2026 https:\\/\\/t.co\\/DQ5uTFnRgW","82":"Tomorrow\\u2019s #watchlist for #NYSE, #NASDAQ between $20 and $30: $ACI $BB $BEN $CLII $CNK $CVM $DBX $FCEL $FOSL $GOVT\\u2026 https:\\/\\/t.co\\/zCWgLYiUzz","83":"Update of my current portfolio:\\n$OPEN\\n$PLTR\\n$CLOV\\n$IPOE\\n$SQ\\n$DMTK\\n$IPOF\\n$CLII\\n$BOX\\nIt is safe to say my portfolio h\\u2026 https:\\/\\/t.co\\/XznciMQz0z","84":"My most bullish argument for $SBE &amp; $CLII is the price of $BLNK, which I\'ve never seen anywhere (working) and I hav\\u2026 https:\\/\\/t.co\\/ts9drD0c9d","85":"$GME currently has a higher market cap than $DKNG \\ud83e\\udd2f","86":"The #NIOET7 has \\u201cself-driving functionality in mind for the future and the sensor package to back it up,\\u201d according\\u2026 https:\\/\\/t.co\\/t9Jf3bwHSf","87":"$DKNG - acting really well today in a weak market. Large base - good VCP characteristics.","88":"RESPECT THE REDDIT ARMY BUT THE SHEIK STOCK ARMY IS THE REAL. I SUPLEX THE $GME NOW BUY $INO $SRNE $BB $MAC $IDEX\\u2026 https:\\/\\/t.co\\/g9q6MueHu9","89":"NIO confirms it tweaked GFA law for critical Assembly vote on NI Protocol - cross community consent will not be req\\u2026 https:\\/\\/t.co\\/c7xuUNaKfy","90":"The video, found in an archive of data uploaded to Parler, includes a fresh look at the mob\\u2019s confrontation with Eu\\u2026 https:\\/\\/t.co\\/EagoxrQZq2","91":"dear journal, today was busy:\\n\\n1. bought some YOLO calls on $gme\\n2. announced i\'m running for governor of CA?\\n3. bo\\u2026 https:\\/\\/t.co\\/C88PU4BX3S","92":"BREAKING: HB113 dies on the House floor in 3rd reading. #mtleg #mtpol","93":"Mara, how does this help change the culture or advance the conversation to meaningful solutions. \\n\\nThis type of sta\\u2026 https:\\/\\/t.co\\/pM6X5j9ZyJ","94":"Once again, @eidietrich executes an ambitious data project that - wait for it - helps Montanans become more engaged\\u2026 https:\\/\\/t.co\\/bKeQ3bU9dN","95":"Steve Bannon broadcasts election denialism and apocalyptic calls to action several times a day via Apple\\u2019s podcast\\u2026 https:\\/\\/t.co\\/Wd51fde7C4","96":"All u had to know is the pattern.  Then your process and system has you profit from it. $jmia. https:\\/\\/t.co\\/DDrHaTO4hG","97":"Put your feet up and enjoy the masterpiece I\\u2019m about to paint! Put $1 on me with @DKSportsbook and you could put $2\\u2026 https:\\/\\/t.co\\/1KmPdXb1sr","98":"$CCIV ain\'t hitting $23 right no for no reason....LFG","99":"EVgo, an electric vehicle charging network, is nearing a deal to merge with a SPAC, Climate Change Crisis Real Impa\\u2026 https:\\/\\/t.co\\/iVG5gQGESJ","100":"Seeing $cciv chatter - they are trying to acquire #lucid - not sure if this is confirmed from @LucidMotors - we shall see tomorrow.  #ev"}}'
+
+static_symbols = ['NIO','TTCF','JMIA','MARA','CCIV','NNDM','DKNG','PLTR','CLII','SBE','HUYA','LI']
+
+static_quotes = '{"Symbol":{"0":"PLTR","1":"JMIA","2":"DKNG","3":"SBE","4":"TTCF","5":"HUYA","6":"LI","7":"NIO","8":"NNDM","9":"MARA","10":"CLII","11":"CCIV"},"Last Price":{"0":39.0,"1":57.37,"2":53.29,"3":40.65,"4":24.25,"5":25.84,"6":30.95,"7":57.15,"8":14.5,"9":17.0,"10":20.1,"11":21.15},"Net Change":{"0":3.63,"1":-0.54,"2":-0.77,"3":-0.62,"4":-0.4,"5":-0.92,"6":-1.55,"7":-3.16,"8":-1.09,"9":-1.6,"10":-1.97,"11":-3.29},"Net % Change":{"0":"10.26","1":"-0.93","2":"-1.42","3":"-1.50","4":"-1.62","5":"-3.44","6":"-4.77","7":"-5.24","8":"-6.99","9":"-8.60","10":"-8.93","11":"-13.46"}}'
+
+static_candle = '{"Symbol":{"0":"NIO","1":"NIO","2":"NIO","3":"NIO","4":"NIO","5":"NIO","6":"NIO","7":"NIO","8":"NIO","9":"NIO","10":"NIO","11":"NIO","12":"NIO","13":"NIO","14":"NIO","15":"NIO","16":"NIO","17":"NIO","18":"NIO","19":"NIO","20":"NIO","21":"NIO","22":"NIO","23":"NIO","24":"NIO","25":"NIO","26":"NIO","27":"NIO","28":"NIO","29":"NIO","30":"NIO","31":"NIO","32":"NIO","33":"NIO","34":"NIO","35":"NIO","36":"NIO","37":"NIO","38":"NIO","39":"NIO","40":"NIO","41":"NIO","42":"NIO","43":"NIO","44":"NIO","45":"NIO","46":"NIO","47":"NIO","48":"NIO","49":"NIO","50":"NIO","51":"NIO","52":"NIO","53":"NIO","54":"NIO","55":"NIO","56":"NIO","57":"NIO","58":"NIO","59":"NIO","60":"NIO","61":"NIO","62":"NIO","63":"NIO","64":"NIO","65":"NIO","66":"NIO","67":"NIO","68":"NIO","69":"NIO","70":"NIO","71":"NIO","72":"NIO","73":"NIO","74":"NIO","75":"NIO","76":"NIO","77":"NIO","78":"NIO","79":"NIO","80":"NIO","81":"NIO","82":"NIO","83":"NIO","84":"NIO","85":"NIO","86":"NIO","87":"NIO","88":"NIO","89":"NIO","90":"NIO","91":"NIO","92":"NIO","93":"NIO","94":"NIO","95":"NIO","96":"NIO","97":"NIO","98":"NIO","99":"NIO","100":"NIO","101":"NIO","102":"NIO","103":"NIO","104":"NIO","105":"NIO","106":"NIO","107":"NIO","108":"NIO","109":"NIO","110":"NIO","111":"NIO","112":"NIO","113":"NIO","114":"NIO","115":"NIO","116":"NIO","117":"NIO","118":"NIO","119":"NIO","120":"NIO","121":"NIO","122":"NIO","123":"NIO","124":"NIO","125":"NIO","126":"NIO","127":"NIO","128":"NIO","129":"NIO","130":"NIO","131":"NIO","132":"NIO","133":"NIO","134":"NIO","135":"NIO","136":"NIO","137":"NIO","138":"NIO","139":"NIO","140":"NIO","141":"NIO","142":"NIO","143":"NIO","144":"NIO","145":"NIO","146":"NIO","147":"NIO","148":"NIO","149":"NIO","150":"NIO","151":"NIO","152":"NIO","153":"NIO","154":"NIO","155":"NIO","156":"JMIA","157":"JMIA","158":"JMIA","159":"JMIA","160":"JMIA","161":"JMIA","162":"JMIA","163":"JMIA","164":"JMIA","165":"JMIA","166":"JMIA","167":"JMIA","168":"JMIA","169":"JMIA","170":"JMIA","171":"JMIA","172":"JMIA","173":"JMIA","174":"JMIA","175":"JMIA","176":"JMIA","177":"JMIA","178":"JMIA","179":"JMIA","180":"JMIA","181":"JMIA","182":"JMIA","183":"JMIA","184":"JMIA","185":"JMIA","186":"JMIA","187":"JMIA","188":"JMIA","189":"JMIA","190":"JMIA","191":"JMIA","192":"JMIA","193":"JMIA","194":"JMIA","195":"JMIA","196":"JMIA","197":"JMIA","198":"JMIA","199":"JMIA","200":"JMIA","201":"JMIA","202":"JMIA","203":"JMIA","204":"JMIA","205":"JMIA","206":"JMIA","207":"JMIA","208":"JMIA","209":"JMIA","210":"JMIA","211":"JMIA","212":"JMIA","213":"JMIA","214":"JMIA","215":"JMIA","216":"JMIA","217":"JMIA","218":"JMIA","219":"JMIA","220":"JMIA","221":"JMIA","222":"JMIA","223":"JMIA","224":"JMIA","225":"JMIA","226":"JMIA","227":"JMIA","228":"JMIA","229":"JMIA","230":"JMIA","231":"JMIA","232":"JMIA","233":"JMIA","234":"JMIA","235":"JMIA","236":"JMIA","237":"JMIA","238":"JMIA","239":"JMIA","240":"JMIA","241":"JMIA","242":"JMIA","243":"JMIA","244":"JMIA","245":"JMIA","246":"JMIA","247":"JMIA","248":"JMIA","249":"JMIA","250":"JMIA","251":"JMIA","252":"JMIA","253":"JMIA","254":"JMIA","255":"JMIA","256":"JMIA","257":"JMIA","258":"JMIA","259":"JMIA","260":"JMIA","261":"JMIA","262":"JMIA","263":"JMIA","264":"JMIA","265":"JMIA","266":"JMIA","267":"JMIA","268":"JMIA","269":"JMIA","270":"JMIA","271":"JMIA","272":"JMIA","273":"JMIA","274":"JMIA","275":"JMIA","276":"JMIA","277":"JMIA","278":"JMIA","279":"JMIA","280":"JMIA","281":"JMIA","282":"JMIA","283":"JMIA","284":"JMIA","285":"JMIA","286":"JMIA","287":"JMIA","288":"JMIA","289":"JMIA","290":"JMIA","291":"JMIA","292":"JMIA","293":"JMIA","294":"JMIA","295":"JMIA","296":"JMIA","297":"JMIA","298":"JMIA","299":"JMIA","300":"JMIA","301":"JMIA","302":"JMIA","303":"JMIA","304":"JMIA","305":"JMIA","306":"JMIA","307":"JMIA","308":"JMIA","309":"JMIA","310":"JMIA","311":"JMIA"},"Date":{"0":1611730800000,"1":1611731100000,"2":1611731400000,"3":1611731700000,"4":1611732000000,"5":1611732300000,"6":1611732600000,"7":1611732900000,"8":1611733200000,"9":1611733500000,"10":1611733800000,"11":1611734100000,"12":1611734400000,"13":1611734700000,"14":1611735000000,"15":1611735300000,"16":1611735600000,"17":1611735900000,"18":1611736200000,"19":1611736500000,"20":1611736800000,"21":1611737100000,"22":1611737400000,"23":1611737700000,"24":1611738000000,"25":1611738300000,"26":1611738600000,"27":1611738900000,"28":1611739200000,"29":1611739500000,"30":1611739800000,"31":1611740100000,"32":1611740400000,"33":1611740700000,"34":1611741000000,"35":1611741300000,"36":1611741600000,"37":1611741900000,"38":1611742200000,"39":1611742500000,"40":1611742800000,"41":1611743100000,"42":1611743400000,"43":1611743700000,"44":1611744000000,"45":1611744300000,"46":1611744600000,"47":1611744900000,"48":1611745200000,"49":1611745500000,"50":1611745800000,"51":1611746100000,"52":1611746400000,"53":1611746700000,"54":1611747000000,"55":1611747300000,"56":1611747600000,"57":1611747900000,"58":1611748200000,"59":1611748500000,"60":1611748800000,"61":1611749100000,"62":1611749400000,"63":1611749700000,"64":1611750000000,"65":1611750300000,"66":1611750600000,"67":1611750900000,"68":1611751200000,"69":1611751500000,"70":1611751800000,"71":1611752100000,"72":1611752400000,"73":1611752700000,"74":1611753000000,"75":1611753300000,"76":1611753600000,"77":1611753900000,"78":1611754200000,"79":1611754500000,"80":1611754800000,"81":1611755100000,"82":1611755400000,"83":1611755700000,"84":1611756000000,"85":1611756300000,"86":1611756600000,"87":1611756900000,"88":1611757200000,"89":1611757500000,"90":1611757800000,"91":1611758100000,"92":1611758400000,"93":1611758700000,"94":1611759000000,"95":1611759300000,"96":1611759600000,"97":1611759900000,"98":1611760200000,"99":1611760500000,"100":1611760800000,"101":1611761100000,"102":1611761400000,"103":1611761700000,"104":1611762000000,"105":1611762300000,"106":1611762600000,"107":1611762900000,"108":1611763200000,"109":1611763500000,"110":1611763800000,"111":1611764100000,"112":1611764400000,"113":1611764700000,"114":1611765000000,"115":1611765300000,"116":1611765600000,"117":1611765900000,"118":1611766200000,"119":1611766500000,"120":1611766800000,"121":1611767100000,"122":1611767400000,"123":1611767700000,"124":1611768000000,"125":1611768300000,"126":1611768600000,"127":1611768900000,"128":1611769200000,"129":1611769500000,"130":1611769800000,"131":1611770100000,"132":1611770400000,"133":1611770700000,"134":1611771000000,"135":1611771300000,"136":1611771600000,"137":1611771900000,"138":1611772200000,"139":1611772500000,"140":1611772800000,"141":1611773100000,"142":1611773400000,"143":1611773700000,"144":1611774000000,"145":1611774300000,"146":1611774600000,"147":1611774900000,"148":1611775200000,"149":1611775500000,"150":1611775800000,"151":1611776100000,"152":1611776400000,"153":1611776700000,"154":1611777000000,"155":1611777300000,"156":1611730800000,"157":1611731100000,"158":1611731400000,"159":1611731700000,"160":1611732000000,"161":1611732300000,"162":1611732600000,"163":1611732900000,"164":1611733200000,"165":1611733500000,"166":1611733800000,"167":1611734100000,"168":1611734400000,"169":1611734700000,"170":1611735000000,"171":1611735300000,"172":1611735600000,"173":1611735900000,"174":1611736200000,"175":1611736500000,"176":1611736800000,"177":1611737100000,"178":1611737400000,"179":1611737700000,"180":1611738000000,"181":1611738300000,"182":1611738600000,"183":1611738900000,"184":1611739200000,"185":1611739500000,"186":1611739800000,"187":1611740100000,"188":1611740400000,"189":1611740700000,"190":1611741000000,"191":1611741300000,"192":1611741600000,"193":1611741900000,"194":1611742200000,"195":1611742500000,"196":1611742800000,"197":1611743100000,"198":1611743400000,"199":1611743700000,"200":1611744000000,"201":1611744300000,"202":1611744600000,"203":1611744900000,"204":1611745200000,"205":1611745500000,"206":1611745800000,"207":1611746100000,"208":1611746400000,"209":1611746700000,"210":1611747000000,"211":1611747300000,"212":1611747600000,"213":1611747900000,"214":1611748200000,"215":1611748500000,"216":1611748800000,"217":1611749100000,"218":1611749400000,"219":1611749700000,"220":1611750000000,"221":1611750300000,"222":1611750600000,"223":1611750900000,"224":1611751200000,"225":1611751500000,"226":1611751800000,"227":1611752100000,"228":1611752400000,"229":1611752700000,"230":1611753000000,"231":1611753300000,"232":1611753600000,"233":1611753900000,"234":1611754200000,"235":1611754500000,"236":1611754800000,"237":1611755100000,"238":1611755400000,"239":1611755700000,"240":1611756000000,"241":1611756300000,"242":1611756600000,"243":1611756900000,"244":1611757200000,"245":1611757500000,"246":1611757800000,"247":1611758100000,"248":1611758400000,"249":1611758700000,"250":1611759000000,"251":1611759300000,"252":1611759600000,"253":1611759900000,"254":1611760200000,"255":1611760500000,"256":1611760800000,"257":1611761100000,"258":1611761400000,"259":1611761700000,"260":1611762000000,"261":1611762300000,"262":1611762600000,"263":1611762900000,"264":1611763200000,"265":1611763500000,"266":1611763800000,"267":1611764100000,"268":1611764400000,"269":1611764700000,"270":1611765000000,"271":1611765300000,"272":1611765600000,"273":1611765900000,"274":1611766200000,"275":1611766500000,"276":1611766800000,"277":1611767100000,"278":1611767400000,"279":1611767700000,"280":1611768000000,"281":1611768300000,"282":1611768600000,"283":1611768900000,"284":1611769200000,"285":1611769500000,"286":1611769800000,"287":1611770100000,"288":1611770400000,"289":1611770700000,"290":1611771000000,"291":1611771300000,"292":1611771600000,"293":1611771900000,"294":1611772200000,"295":1611772500000,"296":1611772800000,"297":1611773100000,"298":1611773400000,"299":1611773700000,"300":1611774000000,"301":1611774300000,"302":1611774600000,"303":1611774900000,"304":1611775200000,"305":1611775500000,"306":1611775800000,"307":1611776100000,"308":1611776400000,"309":1611776700000,"310":1611777000000,"311":1611777300000},"Open":{"0":58.3,"1":58.74,"2":58.56,"3":58.54,"4":58.45,"5":58.29,"6":58.28,"7":58.3,"8":58.26,"9":58.15,"10":58.16,"11":58.18,"12":58.79,"13":57.9,"14":57.57,"15":57.701,"16":57.7,"17":57.9,"18":57.75,"19":57.78,"20":57.9,"21":57.8,"22":57.79,"23":58.17,"24":58.21,"25":57.7102,"26":57.7,"27":57.51,"28":57.39,"29":57.11,"30":57.95,"31":58.015,"32":58.82,"33":57.95,"34":57.9351,"35":58.15,"36":58.4016,"37":57.9777,"38":58.1,"39":57.37,"40":57.94,"41":58.445,"42":58.5899,"43":58.88,"44":59.2613,"45":59.16,"46":59.06,"47":59.5401,"48":59.49,"49":59.66,"50":59.56,"51":59.4,"52":59.11,"53":58.96,"54":59.02,"55":58.975,"56":58.78,"57":58.68,"58":58.53,"59":58.7,"60":58.32,"61":58.35,"62":58.435,"63":58.8,"64":58.73,"65":58.66,"66":58.42,"67":58.33,"68":58.15,"69":58.08,"70":58.2927,"71":58.1816,"72":58.3801,"73":58.26,"74":58.08,"75":58.02,"76":58.01,"77":57.93,"78":58.21,"79":58.35,"80":58.27,"81":58.25,"82":58.325,"83":58.37,"84":58.28,"85":58.22,"86":58.13,"87":58.06,"88":57.7699,"89":57.68,"90":57.5999,"91":57.34,"92":57.6169,"93":57.76,"94":57.5,"95":57.1131,"96":56.96,"97":57.005,"98":57.0098,"99":56.945,"100":56.92,"101":56.81,"102":56.84,"103":56.67,"104":56.9762,"105":57.15,"106":57.33,"107":57.245,"108":57.08,"109":57.22,"110":57.15,"111":56.2,"112":56.55,"113":56.41,"114":56.15,"115":56.6,"116":57.09,"117":57.0,"118":56.99,"119":56.91,"120":56.97,"121":57.05,"122":57.05,"123":57.06,"124":57.0,"125":56.96,"126":57.0,"127":57.1064,"128":57.05,"129":56.95,"130":56.89,"131":56.83,"132":56.84,"133":56.76,"134":56.56,"135":56.68,"136":56.62,"137":56.58,"138":57.15,"139":56.47,"140":56.42,"141":56.35,"142":56.32,"143":56.29,"144":57.15,"145":56.18,"146":56.05,"147":56.25,"148":56.02,"149":56.03,"150":56.17,"151":56.3,"152":56.17,"153":56.2,"154":56.23,"155":56.0,"156":57.5,"157":57.38,"158":57.0,"159":56.99,"160":56.47,"161":55.75,"162":55.25,"163":55.28,"164":55.23,"165":55.27,"166":55.21,"167":55.2,"168":57.0,"169":56.25,"170":53.6,"171":52.81,"172":53.7,"173":54.44,"174":54.39,"175":54.73,"176":56.17,"177":56.44,"178":56.63,"179":56.99,"180":57.1,"181":56.3,"182":55.0,"183":55.66,"184":55.52,"185":56.02,"186":56.0,"187":55.9132,"188":56.4599,"189":55.595,"190":56.94,"191":56.49,"192":58.74,"193":59.06,"194":58.8719,"195":58.775,"196":59.2,"197":60.39,"198":61.86,"199":62.26,"200":61.43,"201":60.98,"202":61.54,"203":62.11,"204":62.635,"205":64.19,"206":64.5001,"207":63.87,"208":63.29,"209":63.71,"210":64.72,"211":64.738,"212":64.91,"213":64.756,"214":64.27,"215":64.375,"216":63.52,"217":61.9819,"218":61.744,"219":62.6226,"220":62.0,"221":61.843,"222":61.047,"223":60.675,"224":61.316,"225":61.25,"226":61.27,"227":61.25,"228":61.245,"229":61.06,"230":61.17,"231":61.14,"232":61.35,"233":61.84,"234":62.08,"235":61.66,"236":62.01,"237":61.69,"238":63.17,"239":62.35,"240":62.605,"241":62.74,"242":62.49,"243":62.2721,"244":62.62,"245":62.1688,"246":62.0,"247":61.02,"248":60.99,"249":60.97,"250":60.18,"251":58.7195,"252":58.795,"253":58.87,"254":58.63,"255":58.3868,"256":58.65,"257":58.0201,"258":58.0761,"259":57.615,"260":57.89,"261":57.525,"262":57.7,"263":57.52,"264":57.37,"265":57.45,"266":57.37,"267":56.4,"268":56.4,"269":57.01,"270":57.02,"271":58.2,"272":58.75,"273":58.68,"274":58.71,"275":58.5,"276":58.33,"277":57.6,"278":58.01,"279":58.64,"280":58.65,"281":58.5,"282":58.68,"283":58.58,"284":58.3,"285":57.92,"286":57.92,"287":57.5,"288":57.5,"289":57.5,"290":57.0,"291":56.7,"292":56.72,"293":56.95,"294":57.37,"295":56.75,"296":56.5,"297":56.5,"298":56.5,"299":56.32,"300":57.37,"301":54.99,"302":54.8,"303":54.5,"304":54.25,"305":54.5,"306":54.11,"307":54.8,"308":55.5,"309":55.0,"310":55.75,"311":55.63},"Close":{"0":58.74,"1":58.56,"2":58.54,"3":58.44,"4":58.3,"5":58.3,"6":58.27,"7":58.25,"8":58.14,"9":58.16,"10":58.18,"11":58.01,"12":57.95,"13":57.57,"14":57.7001,"15":57.7,"16":57.9,"17":57.75,"18":57.78,"19":57.91,"20":57.79,"21":57.7999,"22":58.19,"23":58.2,"24":57.71,"25":57.69,"26":57.5002,"27":57.39,"28":57.12,"29":57.86,"30":58.02,"31":58.8343,"32":57.9688,"33":57.9312,"34":58.1506,"35":58.41,"36":57.99,"37":58.1,"38":57.37,"39":57.95,"40":58.4401,"41":58.5807,"42":58.8801,"43":59.269,"44":59.17,"45":59.0511,"46":59.543,"47":59.4816,"48":59.6798,"49":59.56,"50":59.4,"51":59.12,"52":58.9431,"53":59.03,"54":58.98,"55":58.794,"56":58.68,"57":58.535,"58":58.689,"59":58.326,"60":58.345,"61":58.45,"62":58.79,"63":58.7331,"64":58.665,"65":58.42,"66":58.33,"67":58.1599,"68":58.09,"69":58.295,"70":58.18,"71":58.38,"72":58.25,"73":58.089,"74":58.0199,"75":58.0012,"76":57.93,"77":58.2031,"78":58.3414,"79":58.275,"80":58.25,"81":58.325,"82":58.375,"83":58.3,"84":58.22,"85":58.1124,"86":58.06,"87":57.761,"88":57.69,"89":57.5931,"90":57.34,"91":57.6287,"92":57.7701,"93":57.4885,"94":57.11,"95":56.96,"96":57.0,"97":57.0,"98":56.9416,"99":56.92,"100":56.829,"101":56.8499,"102":56.68,"103":56.97,"104":57.17,"105":57.342,"106":57.24,"107":57.08,"108":57.25,"109":56.53,"110":56.18,"111":56.56,"112":56.4,"113":56.15,"114":56.62,"115":56.98,"116":57.15,"117":57.04,"118":56.91,"119":56.97,"120":57.06,"121":57.05,"122":57.06,"123":57.0,"124":56.96,"125":57.0,"126":57.1,"127":57.05,"128":56.95,"129":56.89,"130":56.85,"131":56.84,"132":56.76,"133":56.6,"134":56.61,"135":56.6,"136":56.58,"137":56.52,"138":56.42,"139":56.4,"140":56.35,"141":56.4,"142":56.3,"143":56.25,"144":56.1,"145":56.05,"146":56.18,"147":56.01,"148":56.03,"149":56.19,"150":56.3,"151":56.17,"152":56.2,"153":56.2,"154":56.01,"155":56.19,"156":57.45,"157":57.0,"158":57.0,"159":56.47,"160":55.8,"161":55.0,"162":55.45,"163":55.23,"164":55.25,"165":55.21,"166":55.2,"167":54.41,"168":54.34,"169":53.52,"170":52.58,"171":53.7,"172":54.43,"173":54.14,"174":54.5001,"175":56.43,"176":56.4,"177":56.6,"178":57.0,"179":57.05,"180":56.3,"181":55.05,"182":55.55,"183":55.52,"184":56.5,"185":56.15,"186":55.92,"187":56.41,"188":55.505,"189":56.83,"190":56.26,"191":58.74,"192":59.01,"193":58.97,"194":58.7705,"195":59.2774,"196":60.31,"197":61.75,"198":62.34,"199":61.44,"200":61.025,"201":61.6,"202":62.3,"203":62.635,"204":64.1692,"205":64.67,"206":63.92,"207":63.28,"208":63.62,"209":64.7376,"210":64.6906,"211":64.79,"212":64.8,"213":64.2856,"214":64.43,"215":63.4,"216":61.6535,"217":61.81,"218":62.6221,"219":62.1199,"220":61.81,"221":61.1,"222":60.54,"223":61.175,"224":61.2667,"225":61.19,"226":61.37,"227":61.3799,"228":61.12,"229":61.26,"230":61.07,"231":61.32,"232":61.79,"233":62.2099,"234":61.5624,"235":62.01,"236":61.745,"237":63.17,"238":62.3933,"239":62.7599,"240":62.7415,"241":62.47,"242":62.49,"243":62.57,"244":62.0199,"245":61.99,"246":61.1,"247":61.0695,"248":60.7901,"249":60.1801,"250":58.72,"251":58.795,"252":58.85,"253":58.56,"254":58.3768,"255":58.6601,"256":58.08,"257":57.98,"258":57.54,"259":57.76,"260":57.73,"261":57.8599,"262":57.69,"263":57.37,"264":57.45,"265":56.5,"266":56.4,"267":56.4,"268":57.01,"269":57.01,"270":57.29,"271":58.72,"272":58.81,"273":58.71,"274":58.5,"275":58.33,"276":57.6,"277":58.67,"278":58.65,"279":58.5,"280":58.5,"281":58.25,"282":58.26,"283":58.49,"284":58.0,"285":57.92,"286":57.5,"287":57.5,"288":57.5,"289":57.1,"290":56.7,"291":56.7,"292":56.89,"293":56.85,"294":56.79,"295":56.5,"296":56.5,"297":56.6,"298":56.8,"299":55.68,"300":55.0,"301":54.92,"302":54.5,"303":54.0,"304":54.4,"305":54.4,"306":54.4,"307":54.96,"308":55.0,"309":55.89,"310":55.63,"311":55.58},"High":{"0":59.05,"1":58.79,"2":58.65,"3":58.55,"4":58.45,"5":58.34,"6":58.42,"7":58.35,"8":58.31,"9":58.24,"10":58.27,"11":58.25,"12":59.0001,"13":57.9002,"14":57.73,"15":57.77,"16":57.95,"17":57.93,"18":57.8,"19":57.94,"20":57.92,"21":57.8,"22":58.19,"23":58.34,"24":58.5,"25":57.85,"26":57.71,"27":57.64,"28":57.4002,"29":58.05,"30":58.38,"31":58.92,"32":58.9399,"33":58.24,"34":58.3498,"35":58.5,"36":58.45,"37":58.19,"38":58.1,"39":58.03,"40":58.45,"41":58.77,"42":59.01,"43":59.2899,"44":59.37,"45":59.177,"46":59.75,"47":59.68,"48":59.698,"49":59.79,"50":59.685,"51":59.46,"52":59.18,"53":59.17,"54":59.22,"55":59.16,"56":58.84,"57":58.85,"58":58.74,"59":58.77,"60":58.4,"61":58.55,"62":58.83,"63":58.89,"64":58.82,"65":58.67,"66":58.43,"67":58.3799,"68":58.24,"69":58.33,"70":58.3016,"71":58.4003,"72":58.39,"73":58.26,"74":58.12,"75":58.07,"76":58.01,"77":58.35,"78":58.3769,"79":58.38,"80":58.41,"81":58.35,"82":58.55,"83":58.48,"84":58.37,"85":58.23,"86":58.165,"87":58.09,"88":57.81,"89":57.755,"90":57.65,"91":57.6699,"92":57.8,"93":57.76,"94":57.54,"95":57.22,"96":57.09,"97":57.23,"98":57.05,"99":57.08,"100":57.12,"101":56.965,"102":56.96,"103":56.97,"104":57.21,"105":57.4,"106":57.37,"107":57.34,"108":57.35,"109":57.35,"110":57.15,"111":56.58,"112":57.0,"113":56.51,"114":56.65,"115":57.15,"116":57.26,"117":58.0,"118":57.1,"119":56.99,"120":57.06,"121":57.07,"122":57.07,"123":57.16,"124":57.07,"125":57.04,"126":57.15,"127":57.15,"128":57.05,"129":57.02,"130":57.0,"131":56.92,"132":56.9,"133":56.76,"134":56.7,"135":56.7,"136":56.66,"137":56.64,"138":57.15,"139":57.0,"140":56.45,"141":56.45,"142":56.43,"143":56.29,"144":57.15,"145":56.18,"146":56.25,"147":56.25,"148":56.05,"149":56.2,"150":56.3,"151":56.3,"152":56.2,"153":56.25,"154":56.24,"155":56.25,"156":57.68,"157":57.38,"158":57.43,"159":56.99,"160":56.47,"161":55.8,"162":55.88,"163":55.65,"164":55.63,"165":55.31,"166":55.65,"167":55.2,"168":57.5,"169":56.25,"170":53.87,"171":54.0,"172":54.45,"173":54.5,"174":55.0,"175":56.5,"176":56.56,"177":56.86,"178":58.0,"179":57.2,"180":57.1,"181":56.3,"182":55.68,"183":55.77,"184":57.68,"185":56.55,"186":56.2,"187":57.0,"188":56.6439,"189":56.9,"190":57.2593,"191":58.74,"192":59.3233,"193":59.09,"194":59.5499,"195":59.2999,"196":60.48,"197":62.32,"198":62.67,"199":62.47,"200":61.65,"201":61.6,"202":62.5299,"203":62.83,"204":64.2,"205":65.47,"206":64.625,"207":64.27,"208":63.917,"209":64.75,"210":64.97,"211":65.5,"212":65.18,"213":65.0,"214":64.88,"215":64.59,"216":63.591,"217":62.4659,"218":62.9007,"219":62.81,"220":62.28,"221":61.87,"222":61.29,"223":61.46,"224":61.7985,"225":61.62,"226":61.55,"227":61.66,"228":61.36,"229":61.26,"230":61.3,"231":61.49,"232":62.05,"233":62.3177,"234":62.46,"235":62.26,"236":62.08,"237":63.3875,"238":63.2218,"239":62.7718,"240":62.98,"241":62.86,"242":62.8225,"243":62.99,"244":62.68,"245":62.19,"246":62.04,"247":61.105,"248":61.24,"249":61.07,"250":60.4391,"251":59.6791,"252":58.86,"253":59.1669,"254":58.7,"255":59.0899,"256":58.9662,"257":58.45,"258":58.27,"259":57.92,"260":57.97,"261":58.38,"262":58.07,"263":58.0,"264":57.94,"265":57.92,"266":57.37,"267":56.69,"268":57.94,"269":57.8,"270":58.34,"271":59.85,"272":59.78,"273":58.81,"274":58.71,"275":58.71,"276":58.5,"277":58.88,"278":58.74,"279":58.74,"280":58.77,"281":58.88,"282":58.7,"283":58.59,"284":58.45,"285":58.15,"286":58.1,"287":57.8,"288":57.54,"289":57.5,"290":57.1,"291":57.09,"292":57.1,"293":56.95,"294":57.37,"295":56.75,"296":56.9,"297":56.79,"298":56.83,"299":56.41,"300":57.37,"301":55.0,"302":54.98,"303":55.01,"304":54.96,"305":54.8,"306":54.53,"307":55.52,"308":55.8,"309":55.9,"310":55.88,"311":56.88},"Low":{"0":58.25,"1":58.5,"2":58.5,"3":58.42,"4":58.29,"5":58.22,"6":58.25,"7":58.25,"8":58.14,"9":58.08,"10":58.15,"11":58.0,"12":57.8798,"13":57.52,"14":57.45,"15":57.64,"16":57.68,"17":57.73,"18":57.74,"19":57.76,"20":57.79,"21":57.77,"22":57.79,"23":58.17,"24":57.7,"25":57.69,"26":57.5,"27":57.38,"28":57.11,"29":57.1,"30":57.8401,"31":57.31,"32":57.91,"33":57.84,"34":57.785,"35":57.9,"36":57.735,"37":57.72,"38":57.36,"39":57.22,"40":57.8,"41":58.43,"42":58.526,"43":58.8327,"44":59.05,"45":58.85,"46":59.0331,"47":59.43,"48":59.36,"49":59.4931,"50":59.3,"51":58.985,"52":58.87,"53":58.9,"54":58.971,"55":58.77,"56":58.68,"57":58.48,"58":58.48,"59":58.3,"60":58.17,"61":58.35,"62":58.41,"63":58.71,"64":58.63,"65":58.32,"66":58.23,"67":58.01,"68":58.03,"69":58.0,"70":58.13,"71":58.1601,"72":58.24,"73":58.0,"74":57.99,"75":57.99,"76":57.73,"77":57.91,"78":58.15,"79":58.2,"80":58.18,"81":58.16,"82":58.31,"83":58.25,"84":58.18,"85":58.02,"86":57.97,"87":57.75,"88":57.52,"89":57.55,"90":57.26,"91":57.24,"92":57.511,"93":57.4731,"94":57.0899,"95":56.74,"96":56.61,"97":56.9,"98":56.71,"99":56.85,"100":56.81,"101":56.6,"102":56.65,"103":56.65,"104":56.921,"105":57.06,"106":57.03,"107":57.0,"108":57.08,"109":56.53,"110":55.92,"111":56.08,"112":56.34,"113":56.07,"114":56.15,"115":56.58,"116":56.95,"117":56.7,"118":56.85,"119":56.75,"120":56.91,"121":56.89,"122":57.0,"123":56.9499,"124":56.92,"125":56.95,"126":56.95,"127":57.02,"128":56.87,"129":56.88,"130":56.82,"131":56.78,"132":56.7,"133":56.55,"134":56.51,"135":56.58,"136":56.5,"137":56.5,"138":56.42,"139":56.4,"140":56.3,"141":56.3,"142":56.29,"143":56.15,"144":56.1,"145":56.0,"146":56.0,"147":56.01,"148":56.0,"149":56.0,"150":56.15,"151":56.15,"152":56.15,"153":56.15,"154":56.0,"155":55.96,"156":56.71,"157":57.0,"158":57.0,"159":56.32,"160":55.7,"161":55.0,"162":54.56,"163":55.1,"164":55.2,"165":55.15,"166":55.16,"167":54.06,"168":53.85,"169":53.52,"170":52.58,"171":52.7,"172":53.7,"173":53.7,"174":54.12,"175":54.51,"176":56.17,"177":56.4,"178":56.6,"179":56.6,"180":56.0,"181":55.0,"182":55.0,"183":55.3,"184":55.3,"185":56.0,"186":54.25,"187":54.5001,"188":55.37,"189":55.21,"190":55.69,"191":56.17,"192":58.57,"193":58.2426,"194":58.6,"195":58.23,"196":58.9601,"197":60.35,"198":61.3206,"199":61.32,"200":60.51,"201":60.67,"202":61.538,"203":62.05,"204":62.6,"205":64.01,"206":63.8,"207":63.05,"208":63.09,"209":63.53,"210":64.3,"211":64.2,"212":64.12,"213":63.93,"214":63.9,"215":63.38,"216":61.22,"217":61.5,"218":61.62,"219":62.0,"220":61.6,"221":60.8803,"222":60.25,"223":60.13,"224":61.11,"225":60.83,"226":60.1,"227":61.1,"228":60.93,"229":60.77,"230":60.87,"231":61.0,"232":61.34,"233":61.7816,"234":61.5,"235":61.58,"236":61.68,"237":61.57,"238":62.14,"239":62.152,"240":62.46,"241":62.2,"242":62.2101,"243":62.2721,"244":61.8,"245":61.71,"246":61.02,"247":60.5677,"248":60.7069,"249":60.1701,"250":58.3201,"251":58.41,"252":57.75,"253":58.56,"254":57.8,"255":58.21,"256":58.01,"257":57.75,"258":57.5,"259":57.15,"260":57.51,"261":57.14,"262":57.33,"263":57.32,"264":57.32,"265":56.11,"266":56.0,"267":56.0,"268":56.3,"269":57.0,"270":57.0,"271":57.92,"272":57.37,"273":57.37,"274":58.3,"275":58.3,"276":56.0,"277":57.6,"278":57.93,"279":57.37,"280":58.08,"281":58.12,"282":58.26,"283":58.3,"284":58.0,"285":57.62,"286":57.5,"287":57.4,"288":57.37,"289":57.0,"290":56.7,"291":56.68,"292":56.7,"293":56.73,"294":56.68,"295":56.25,"296":56.5,"297":56.5,"298":56.12,"299":55.68,"300":54.51,"301":54.5,"302":54.38,"303":54.0,"304":54.01,"305":54.13,"306":54.01,"307":54.8,"308":55.0,"309":54.81,"310":55.5,"311":55.58}}'
